競合を解消
