+++ app/build.gradle
@@ -31,9 +31,9 @@ android {
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile "com.android.support:support-v4:19.0.+"
-    compile 'com.fasterxml.jackson.core:jackson-core:2.5.+'
-    compile 'com.fasterxml.jackson.core:jackson-databind:2.5.+'
+    compile "com.android.support:support-v4:+"
+    compile 'com.fasterxml.jackson.core:jackson-core:2.5.4'
+    compile 'com.fasterxml.jackson.core:jackson-databind:2.5.4'
 
     repositories {
         mavenCentral()
