+++ app/src/androidTest/java/jp/bunkatusoft/explorersofsettlement/screen/world/WorldFieldUtilTest.java
@@ -1,4 +1,4 @@
-package jp.bunkatusoft.explorersofsettlement.field.world;
+package jp.bunkatusoft.explorersofsettlement.screen.world;
 
 import android.support.test.runner.AndroidJUnit4;
 import android.test.suitebuilder.annotation.MediumTest;
