+++ app/src/main/java/jp/bunkatusoft/explorersofsettlement/Main.java
@@ -4,7 +4,7 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.support.v4.app.FragmentActivity;
 
-import jp.bunkatusoft.explorersofsettlement.title.TitleActivity;
+import jp.bunkatusoft.explorersofsettlement.screen.logo.LogoActivity;
 
 
 public class Main extends FragmentActivity {
@@ -14,7 +14,7 @@ public class Main extends FragmentActivity {
 		super.onCreate(savedInstanceState);
 		//requestWindowFeature(Window.FEATURE_NO_TITLE);
 		setContentView(R.layout.activity_main);
-		startActivity(new Intent(this, TitleActivity.class));
+		startActivity(new Intent(this, LogoActivity.class));
 		finish();
 	}
