+++ app/src/main/java/jp/bunkatusoft/explorersofsettlement/screen/world/DynamicCommandGroup.java
@@ -1,4 +1,4 @@
-package jp.bunkatusoft.explorersofsettlement.field.world;
+package jp.bunkatusoft.explorersofsettlement.screen.world;
 
 import android.content.Context;
 import android.view.Gravity;
@@ -8,36 +8,41 @@ import android.view.animation.AnimationSet;
 import android.widget.ImageButton;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
-import android.widget.LinearLayout.LayoutParams;
-import android.widget.RelativeLayout;
+
+import java.util.List;
 
 import jp.bunkatusoft.explorersofsettlement.R;
 import jp.bunkatusoft.explorersofsettlement.system.LinearButtonCommandGroup;
 
-public class StaticCommandGroup extends LinearButtonCommandGroup{
+public class DynamicCommandGroup extends LinearButtonCommandGroup {
 
-	interface OnStaticCommandClickListener{
-		void OnStaticCommandClick(StaticCommandEnum command);
+	interface OnDynamicCommandClickListener{
+		void OnDynamicCommandClick(DynamicCommandEnum command);
 	}
 
-	OnStaticCommandClickListener mListener;
+	OnDynamicCommandClickListener mListener;
 
-	public StaticCommandGroup(Context context, RelativeLayout rootLayout, OnStaticCommandClickListener listener) {
+	public DynamicCommandGroup(Context context, ViewGroup rootLayout, OnDynamicCommandClickListener listener) {
 		super(context, rootLayout);
 
 		mListener = listener;
 
-		addCommandButton(StaticCommandEnum.PARTY);
-		addCommandButton(StaticCommandEnum.INVENTORY);
-		addCommandButton(StaticCommandEnum.MOVE);
+
+	}
+
+	public void setCommandList(List<DynamicCommandEnum> useCommandList) {
+		mBaseLayout.removeAllViews();
+		if(useCommandList != null && useCommandList.size()>0){
+			for(DynamicCommandEnum command : useCommandList){
+				addCommandButton(command);
+			}
+		}
 	}
 
 	@Override
 	protected void initBaseLayout(ViewGroup rootLayout) {
 		mBaseLayout = new LinearLayout(mContext);
 		mBaseLayout.setOrientation(LinearLayout.HORIZONTAL);
-
-		//実装方法の関係上、rootLayoutはここでは使用しない
 	}
 
 	@Override
@@ -45,21 +50,21 @@ public class StaticCommandGroup extends LinearButtonCommandGroup{
 		return mBaseLayout;
 	}
 
-	protected void addCommandButton(StaticCommandEnum command) {
+	protected void addCommandButton(DynamicCommandEnum command) {
 		ImageButton imageButton = new ImageButton(mContext);
 		imageButton.setImageResource(command.getCommandResourceId());
 		imageButton.setBackgroundResource(R.drawable.icon_background);
 		imageButton.setScaleType(ImageView.ScaleType.CENTER_INSIDE);
 
-		final StaticCommandEnum sendCommand = command;
+		final DynamicCommandEnum sendCommand = command;
 		imageButton.setOnClickListener(new View.OnClickListener() {
 			@Override
 			public void onClick(View view) {
-				mListener.OnStaticCommandClick(sendCommand);
+				mListener.OnDynamicCommandClick(sendCommand);
 			}
 		});
 
-		LayoutParams layoutParams = new LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT);
+		LinearLayout.LayoutParams layoutParams = new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT);
 		layoutParams.gravity = Gravity.CENTER;
 		layoutParams.setMargins(4,4,4,4);
