+++ app/src/main/java/jp/bunkatusoft/explorersofsettlement/screen/world/PieceTypeEnum.java
@@ -1,4 +1,4 @@
-package jp.bunkatusoft.explorersofsettlement.field.world;
+package jp.bunkatusoft.explorersofsettlement.screen.world;
 
 import java.util.NoSuchElementException;
