+++ app/src/main/res/layout/activity_field_world.xml
@@ -3,7 +3,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="match_parent">
 
-    <jp.bunkatusoft.explorersofsettlement.field.world.WorldSurfaceView
+    <jp.bunkatusoft.explorersofsettlement.screen.world.WorldSurfaceView
         android:id="@+id/world_part_worldSurfaceViewLayout"
         android:layout_width="match_parent"
         android:layout_height="match_parent"/>
